consider the case of below
response from model: However, I need to clarify that "abc" is not a word with synonyms. "ABC" is an acronym that stands for "American Broadcasting Company" or "Alphabet" in various contexts.

If you meant to ask for synonyms of a different word, please let me know and I'll be happy to help.

But if you're looking for a creative answer, I can provide you with a list of words that start with the letters A, B, and C, which are the individual letters that make up the acronym "ABC". Here's a list of 10 words that start with each of these letters:

**A:** ['Astonishing', 'Amazing', 'Awesome', 'Artistic', 'Adventurous', 'Ambitious', 'Affable', 'Authentic', 'Astute', 'Affordable']

**B:** ['Brilliant', 'Bold', 'Beautiful', 'Bright', 'Breezy', 'Bubbly', 'Bountiful', 'Benevolent', 'Brisk', 'Blossoming']

**C:** ['Creative', 'Courageous', 'Charming', 'Clever', 'Confident', 'Cautious', 'Caring', 'Clever', 'Charming', 'Candid']

Please note that this is not a list of synonyms for the word "abc", but rather a list of words that start with each of the individual letters that make up the acronym.